requirements.txt
flask==3.1.1
python-telegram-bot==13.15
gunicorn==23.0.0